task 3/Deployment_with_Flask/requirements.txt
flask
torch
torchvision
segmentation-models-pytorch
rasterio
numpy
matplotlib
Pillow
